robot_control_system/requirements.txt
pyserial
flask
pyyaml
numpy
# opencv-python-headless: install via apt instead — no armv7l wheel on piwheels
#   sudo apt install -y python3-opencv

# Raspberry Pi camera only — install via apt, not pip:
#   sudo apt install -y python3-picamera2
